requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels>=4.0.0
channels_redis>=4.1.0
constantly==23.10.4
cryptography==45.0.4
daphne>=4.0.0
Django>=4.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
redis>=5.0.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
zope.interface==7.2
google-cloud-vision>=3.4.0
google-cloud-language>=2.11.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
